Move index.ts
